requirements.txt
flask
PyYAML
google-cloud-storage
